Assignment 2- Construct-AI-QR Code-Generator/requirements.txt
qrcode[pil]==8.0
Pillow==11.1.0
